backend/requirements.txt
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fsspec==2023.10.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
openai-whisper==20230918
packaging==23.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
sentencepiece==0.1.99
sympy==1.12
tiktoken==0.3.3
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.17.0
